pyproject.toml
[tool.poetry]
name = "intellius"
version = "0.1.0"
description = ""
authors = ["김샬롬 <johnkdo2020@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.13"
uvicorn = ">=0.37.0,<0.38.0"
fastapi = ">=0.118.0,<0.119.0"
pydantic = ">=2.11.9,<3.0.0"
sqlmodel = ">=0.0.25,<0.0.26"
pydantic-settings = ">=2.11.0,<3.0.0"
sqlalchemy = ">=2.0.43,<3.0.0"
aiomysql = ">=0.2.0,<0.3.0"
python-jose = {extras = ["cryptography"], version = ">=3.3.0,<4.0.0"}
passlib = {extras = ["bcrypt"], version = ">=1.7.4,<2.0.0"}
email-validator = ">=2.0.0,<3.0.0"
redis = ">=5.0.0,<6.0.0"
gunicorn = ">=23.0.0,<24.0.0"
boto3 = ">=1.35.0,<2.0.0"
botocore = ">=1.35.0,<2.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"